+++ src/Utils.cpp
@@ -60,7 +60,7 @@ std::string convertToScriptPath(const std::string& path)
     for (auto it = p.begin(); it != p.end(); ++it)
     {
         auto str = it->string();
-        if (str.find(' ') != std::string::npos)
+        if (!Luau::isIdentifier(str))
             output += "[\"" + str + "\"]";
         else if (str == ".")
         {